chore: rename kms to aws-kms
