+++ src/crypto/multiple.rs
@@ -5,8 +5,8 @@ use futures_util::Future;
 use crate::error::{self, KmsError};
 
 pub enum Multiple {
-    #[cfg(feature = "kms")]
-    AwsKms(super::kms::KmsClient),
+    #[cfg(feature = "aws-kms")]
+    AwsKms(super::aws_kms::KmsClient),
     #[cfg(feature = "hashicorp-vault")]
     VaultKv2(super::hcvault::HashiCorpVault<super::hcvault::Kv2>),
     None(super::hollow::NoEncryption),
@@ -18,9 +18,9 @@ impl Multiple {
     ) -> error_stack::Result<Self, error::ConfigurationError> {
         #[allow(unreachable_patterns)]
         match config {
-            #[cfg(feature = "kms")]
+            #[cfg(feature = "aws-kms")]
             Some(crate::config::EncryptionScheme::AwsKms(config)) => {
-                Ok(Self::AwsKms(super::kms::KmsClient::new(config).await))
+                Ok(Self::AwsKms(super::aws_kms::KmsClient::new(config).await))
             }
             #[cfg(feature = "hashicorp-vault")]
             Some(crate::config::EncryptionScheme::VaultKv2(config)) => {
@@ -37,7 +37,7 @@ impl<I: super::FromEncoded> super::Decode<I, String> for Multiple {
 
     fn decode(&self, input: String) -> Self::ReturnType<'_, I> {
         match self {
-            #[cfg(feature = "kms")]
+            #[cfg(feature = "aws-kms")]
             Self::AwsKms(client) => client.decode(input),
             #[cfg(feature = "hashicorp-vault")]
             Self::VaultKv2(client) => client.decode(input),
